+++ lib/src/generators/kingdoms/kingdom_generator.dart
@@ -1,5 +1,6 @@
 import '../../entities/kingdoms/kingdom.dart';
 import '../../entities/npcs/npc.dart';
+import '../../subtypes/holidays/holiday_type.dart';
 import '../../subtypes/kingdoms/government_types/government_type.dart';
 import '../../subtypes/kingdoms/kingdom_type.dart';
 import '../../subtypes/races/race.dart';
@@ -7,11 +8,14 @@ import '../base/batch_generator.dart';
 import '../base/constant_generator.dart ';
 import '../base/future_generator.dart';
 import '../base/generator.dart';
+import '../base/list_batch_generator.dart';
 import '../base/number_generator.dart';
+import '../base/repeated_generator.dart';
 import '../base/seed_generator.dart';
 import '../base/unique_generator.dart';
 import '../emblems/emblem_generator.dart';
 import '../fixable.dart';
+import '../holidays/holiday_generator.dart';
 import '../npcs/npc_generator.dart';
 import 'guilds/kingdom_guild_generator.dart';
 import 'settlements/kingdom_settlement_generator.dart';
@@ -25,6 +29,8 @@ class KingdomGenerator implements Generator<Kingdom> {
   static const _numberOfSettlements = 3;
   static const _minNumberOfGuilds = 1;
   static const _maxNumberOfGuilds = 2;
+  static const _minNumberOfHolidays = 1;
+  static const _maxNumberOfHolidays = 3;
 
   KingdomGenerator(this._kingdomType, this._race, this._governmentType) {
     _seed = SeedGenerator.generate();
@@ -45,13 +51,24 @@ class KingdomGenerator implements Generator<Kingdom> {
     numberOfGuildsGenerator.seed((_seed + 3) % SeedGenerator.maxSeed);
     final numberOfGuilds = numberOfGuildsGenerator.generate();
 
+    final numberOfHolidaysGenerator =
+        NumberGenerator(_minNumberOfHolidays, _maxNumberOfHolidays)
+          ..seed((_seed + 4) % SeedGenerator.maxSeed);
+    final holidayTypesGenerator = RepeatedGenerator(
+      _kingdomType.getHolidayTypeGenerator(),
+      numberOfHolidaysGenerator.generate(),
+    )..seed((_seed + 5) % SeedGenerator.maxSeed);
+    final holidayTypes = holidayTypesGenerator.generate();
+
     final generator = BatchGenerator(_getBatch(
       numberOfLeaders,
       numberOfGuilds,
       _governmentType,
       name,
+      holidayTypes,
     ));
-    generator.seed((_seed + 4) % SeedGenerator.maxSeed);
+
+    generator.seed((_seed + 6) % SeedGenerator.maxSeed);
     Kingdom kingdom = Kingdom.fromMap(generator.generate());
 
     if (_kingdomType is Fixable<Kingdom>) {
@@ -66,6 +83,7 @@ class KingdomGenerator implements Generator<Kingdom> {
     int numberOfGuilds,
     GovernmentType governmentType,
     String kingdomName,
+    List<HolidayType> holidayTypes,
   ) =>
       {
         "name": ConstantGenerator(kingdomName),
@@ -106,6 +124,12 @@ class KingdomGenerator implements Generator<Kingdom> {
           (guilds) => guilds.map((e) => e.toMap()).toList(),
         ),
         "trouble": _kingdomType.getTroubleGenerator(),
+        "holidays": FutureGenerator(
+          ListBatchGenerator(
+            holidayTypes.map((x) => HolidayGenerator(x)).toList(),
+          ),
+          (holidays) => holidays.map((x) => x.toMap()).toList(),
+        ),
       };
 
   Generator<List<Npc>> _getLeadersGenerator(